#23-add toggle narbar and fix struc component successfully
