+++ src/assets/css/home.css
@@ -9,18 +9,39 @@
     margin-bottom: 10px;
 }
 
-.card.shadow.border-0:hover {
-    transform: rotate(10deg);
-    background-color: rgb(139, 136, 140);
+
+.card.shadow.border-0 {
+    opacity: 1;
+    transition: opacity 0.6s ease;
+}
+
+.card-container:hover .card.shadow.border-0:not(:hover) {
+    opacity: 0.4;
+}
+
+.card-container:hover .card.shadow.border-0:hover {
+    opacity: 1;
+    background-color: rgba(43, 149, 215, 0.433);
 }
 
 .carousel-item img {
+    border-radius: 15px;
     background-repeat: no-repeat;
     background-size: cover;
     width: 100%;
     height: 400px;
 }
 
-.h2.mb-0.ls-tight{
+.h2.mb-0.ls-tight {
     font-family: 'Helvetica', Courier, monospace;
+}
+
+.btnToggleNavBar {
+    position: relative;
+    top: 200px;
+    width: 40px;
+    height: 10px;
+    right: 30px;
+    z-index: 9999;
+    margin: 0;
 }
