+++ src/pages/ClassRoom/ClassRoomContent.jsx
@@ -1,0 +1,139 @@
+import React from 'react';
+
+function ClassRoomContent() {
+    return (
+        <main className="py-6 bg-surface-secondary">
+            <div className="container">
+                <div className=" row">
+                    <div className="col-4 mb-5">
+                        <div class="card border border-black" >
+                            <div class="card-body">
+                                <div class="card-title classroom ps-2 pt-1 mb-2" style={{ backgroundColor: `#${Math.floor(Math.random() * 16777215).toString(16)}` }}>
+                                    <h5><a href="" className="nameofSubject text-black">Lớp lập trình hướng đối tượng</a></h5>
+                                    <p>Object - Orented Programing</p>
+                                    <img src="http://bootdey.com/img/Content/avatar/avatar1.png" alt="ảnh giảng viên..."
+                                        className="rounded-circle position-absolute bottom-30 end-10" width="70" />
+
+                                </div>
+                                <div className="card-text mb-2">
+                                    <p><b>Giáo viên: </b>Nguyễn Văn A</p>
+                                    <p><b>Phòng: </b>A1-101 | <b>Nhóm: </b> 01</p>
+                                </div>
+                                <div className="joinRoomBtn d-flex justify-content-end">
+                                    <a href="/classroom/detail" class="btn btn-success ">Vào phòng học <i class="bi bi-arrow-bar-right"></i></a>
+
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="col-4 mb-5">
+                        <div class="card border border-black" >
+                            <div class="card-body">
+                                <div class="card-title classroom ps-2 pt-1 mb-2" style={{ backgroundColor: `#${Math.floor(Math.random() * 16777215).toString(16)}` }}>
+                                    <h5><a href="" className="nameofSubject text-black">Lớp lập trình hướng đối tượng</a></h5>
+                                    <p>Object - Orented Programing</p>
+                                    <img src="http://bootdey.com/img/Content/avatar/avatar1.png" alt="ảnh giảng viên..."
+                                        className="rounded-circle position-absolute bottom-30 end-10" width="70" />
+
+                                </div>
+                                <div className="card-text mb-2">
+                                    <p><b>Giáo viên: </b>Nguyễn Văn A</p>
+                                    <p><b>Phòng: </b>A1-101 | <b>Nhóm: </b> 01</p>
+                                </div>
+                                <div className="joinRoomBtn d-flex justify-content-end">
+                                    <a href="#" class="btn btn-success ">Vào phòng học <i class="bi bi-arrow-bar-right"></i></a>
+
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="col-4 mb-5">
+                        <div class="card border border-black" >
+                            <div class="card-body">
+                                <div class="card-title classroom ps-2 pt-1 mb-2" style={{ backgroundColor: `#${Math.floor(Math.random() * 16777215).toString(16)}` }}>
+                                    <h5><a href="" className="nameofSubject text-black">Lớp lập trình hướng đối tượng</a></h5>
+                                    <p>Object - Orented Programing</p>
+                                    <img src="http://bootdey.com/img/Content/avatar/avatar1.png" alt="ảnh giảng viên..."
+                                        className="rounded-circle position-absolute bottom-30 end-10" width="70" />
+
+                                </div>
+                                <div className="card-text mb-2">
+                                    <p><b>Giáo viên: </b>Nguyễn Văn A</p>
+                                    <p><b>Phòng: </b>A1-101 | <b>Nhóm: </b> 01</p>
+                                </div>
+                                <div className="joinRoomBtn d-flex justify-content-end">
+                                    <a href="#" class="btn btn-success ">Vào phòng học <i class="bi bi-arrow-bar-right"></i></a>
+
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="col-4 mb-5">
+                        <div class="card border border-black" >
+                            <div class="card-body">
+                                <div class="card-title classroom ps-2 pt-1 mb-2" style={{ backgroundColor: `#${Math.floor(Math.random() * 16777215).toString(16)}` }}>
+                                    <h5><a href="" className="nameofSubject text-black">Lớp lập trình hướng đối tượng</a></h5>
+                                    <p>Object - Orented Programing</p>
+                                    <img src="http://bootdey.com/img/Content/avatar/avatar1.png" alt="ảnh giảng viên..."
+                                        className="rounded-circle position-absolute bottom-30 end-10" width="70" />
+
+                                </div>
+                                <div className="card-text mb-2">
+                                    <p><b>Giáo viên: </b>Nguyễn Văn A</p>
+                                    <p><b>Phòng: </b>A1-101 | <b>Nhóm: </b> 01</p>
+                                </div>
+                                <div className="joinRoomBtn d-flex justify-content-end">
+                                    <a href="#" class="btn btn-success ">Vào phòng học <i class="bi bi-arrow-bar-right"></i></a>
+
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="col-4 mb-5">
+                        <div class="card border border-black" >
+                            <div class="card-body">
+                                <div class="card-title classroom ps-2 pt-1 mb-2" style={{ backgroundColor: `#${Math.floor(Math.random() * 16777215).toString(16)}` }}>
+                                    <h5><a href="" className="nameofSubject text-black">Lớp lập trình hướng đối tượng</a></h5>
+                                    <p>Object - Orented Programing</p>
+                                    <img src="http://bootdey.com/img/Content/avatar/avatar1.png" alt="ảnh giảng viên..."
+                                        className="rounded-circle position-absolute bottom-30 end-10" width="70" />
+
+                                </div>
+                                <div className="card-text mb-2">
+                                    <p><b>Giáo viên: </b>Nguyễn Văn A</p>
+                                    <p><b>Phòng: </b>A1-101 | <b>Nhóm: </b> 01</p>
+                                </div>
+                                <div className="joinRoomBtn d-flex justify-content-end">
+                                    <a href="/classroom/detail" class="btn btn-success ">Vào phòng học <i class="bi bi-arrow-bar-right"></i></a>
+
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="col-4 mb-5">
+                        <div class="card border border-black" >
+                            <div class="card-body">
+                                <div class="card-title classroom ps-2 pt-1 mb-2" style={{ backgroundColor: `#${Math.floor(Math.random() * 16777215).toString(16)}` }}>
+                                    <h5><a href="" className="nameofSubject text-black">Lớp lập trình hướng đối tượng</a></h5>
+                                    <p>Object - Orented Programing</p>
+                                    <img src="http://bootdey.com/img/Content/avatar/avatar1.png" alt="ảnh giảng viên..."
+                                        className="rounded-circle position-absolute bottom-30 end-10" width="70" />
+
+                                </div>
+                                <div className="card-text mb-2">
+                                    <p><b>Giáo viên: </b>Nguyễn Văn A</p>
+                                    <p><b>Phòng: </b>A1-101 | <b>Nhóm: </b> 01</p>
+                                </div>
+                                <div className="joinRoomBtn d-flex justify-content-end ">
+                                    <a href="#" class="btn btn-success">Vào phòng học <i class="bi bi-arrow-bar-right"></i></a>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+
+        </main>
+    );
+}
+export default React.memo(ClassRoomContent);
